// .shopify-cms/types/shopify.ts
export type ShopifySelect = {
  id: string;
  label: string;
  options: { label: string; value: string; group?: string }[];
  type: "select";
  default?: string;
  info?: string;
};